// src/api/generated/models/productAttributesListParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { ProductAttributesListSortBy } from './productAttributesListSortBy';
import type { ProductAttributesListSortOrder } from './productAttributesListSortOrder';
import type { ProductAttributesListType } from './productAttributesListType';
import type { ProductAttributesListUsageStatus } from './productAttributesListUsageStatus';

export type ProductAttributesListParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
search?: string;
usageStatus?: ProductAttributesListUsageStatus;
createdFrom?: string;
createdTo?: string;
sortOrder?: ProductAttributesListSortOrder;
sortBy?: ProductAttributesListSortBy;
type?: ProductAttributesListType;
};